final changes

+++ controllers/AuditingMemberController.php
@@ -35,6 +35,7 @@ class AuditingMemberController extends Controller
      */
     public function actionIndex()
     {
+        if(!Yii::$app->user->isGuest){
         $searchModel = new SearchAuditingMember();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -42,6 +43,9 @@ class AuditingMemberController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }else{
+        throw new \yii\web\ForbiddenHttpException;
+    }
     }
 
     /**
@@ -52,9 +56,13 @@ class AuditingMemberController extends Controller
      */
     public function actionView($id)
     {
+        if(!Yii::$app->user->isGuest){
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
+        }else{
+        throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -64,6 +72,7 @@ class AuditingMemberController extends Controller
      */
     public function actionCreate()
     {
+        if(!Yii::$app->user->isGuest){
         $model = new AuditingMember();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -73,6 +82,9 @@ class AuditingMemberController extends Controller
         return $this->render('create', [
             'model' => $model,
         ]);
+        }else{
+        throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -84,6 +96,7 @@ class AuditingMemberController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(!Yii::$app->user->isGuest){
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -93,6 +106,9 @@ class AuditingMemberController extends Controller
         return $this->render('update', [
             'model' => $model,
         ]);
+        }else{
+        throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -104,9 +120,13 @@ class AuditingMemberController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->isGuest){
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+         }else{
+        throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
