+++ controllers/PaperFacultyController.php
@@ -37,6 +37,7 @@ class PaperFacultyController extends Controller
      */
     public function actionIndex()
     {
+        if(!Yii::$app->user->isGuest){
         $searchModel = new SearchAcademicYear();
         $searchModel = new SearchPaperFaculty();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
@@ -45,6 +46,9 @@ class PaperFacultyController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }else{
+        throw new \yii\web\ForbiddenHttpException;
+    }
     }
 
     /**
@@ -55,9 +59,13 @@ class PaperFacultyController extends Controller
      */
     public function actionView($id)
     {
+        if(!Yii::$app->user->isGuest){
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
+        }else{
+        throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -67,6 +75,7 @@ class PaperFacultyController extends Controller
      */
     public function actionCreate()
     {
+        if(!Yii::$app->user->isGuest){
         $model = new PaperFaculty();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -76,6 +85,9 @@ class PaperFacultyController extends Controller
         return $this->render('create', [
             'model' => $model,
         ]);
+        }else{
+        throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -87,6 +99,7 @@ class PaperFacultyController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(!Yii::$app->user->isGuest){
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -96,6 +109,9 @@ class PaperFacultyController extends Controller
         return $this->render('update', [
             'model' => $model,
         ]);
+        }else{
+        throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -107,9 +123,13 @@ class PaperFacultyController extends Controller
      */
     public function actionDelete($id)
     {
+        if(!Yii::$app->user->isGuest){
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+        }else{
+        throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
